Update changelog and README for version 5.0.0
This commit updates the changelog with the changes introduced in version 5.0.0, namely the introduction of two new service classes, HtmlService and RssService. It also updates the current version number in the README file from 4.0.1 to 5.0.0. The new services enhance the flexibility of game data format handling and improve code organization.